+++ Simulation/src/main/java/org/example/ex2/Ex2Results.java
@@ -7,5 +7,5 @@ import java.util.Map;
 
 public record Ex2Results (
         Ex2Params params,
-        Map<Double, Map<Integer, List<List<Ex2Particle>>>> resultsByKAndW
+        Map<Double, Map<Double, List<List<Ex2Particle>>>> resultsByKAndW
 ) implements Results {}
